Encoding the Original File & Ensuring the ASCII Format
Created a custom .huff binary file, which includes a custom header based on a defined structure (Magic Number, Version, Huffman Tree Size, Last Bit Count, and Original Size), serialized tree, and compressed data. To implement such, write_huff_file() writes the encoded file with the help of serialize_huffman_tree(), serialize_huffman_tree_recursively() (both of which serialize the Huffman Tree in a pre-order traversal), and compress() (which converts the original file into a sequence of bits based on codes from the tree).

Additionally, added checks to ensure an ASCII encoding of an Original File and limited the CHAR_AMOUNT to 128.